Fixed a very nasty bug in reachability
I had this very weird thing where a test would sometimes succeed and sometimes fail and I had no idea what was not determistic about my code so that freaked me out. In the end it turned out that in reachability analysis I threw away everything I knew about previous outcomes as I was not saving that work correctly. The problem was that because of hashing and dictionary shenanigans the order in which the outcomes were processed changed from run to run. With only 1/3 chance did the work I needed run last so that the result would be correct. My god am I happy that I found that bug because I was close to a total breakdown